+++ source/dalicious/math.d
@@ -526,6 +526,135 @@ in (0 <= eps)
 }
 
 
+
+
+/// Convert to given type without errors by bounding values to target type
+/// limits.
+IntTo boundedConvert(IntTo, IntFrom)(IntFrom value) pure nothrow @safe @nogc
+    if (isIntegral!IntTo && isIntegral!IntFrom)
+{
+    static if (isSigned!IntFrom == isSigned!IntTo)
+    {
+        if (IntTo.min <= value && value <= IntTo.max)
+            return cast(IntTo) value;
+        else if (IntTo.min > value)
+            return IntTo.min;
+        else
+            return IntTo.max;
+    }
+    else static if (isSigned!IntFrom)
+    {
+        static assert(isUnsigned!IntTo);
+
+        if (value < 0)
+            return IntTo.min;
+        else if (cast(Unsigned!IntFrom) value < IntTo.max)
+            return cast(IntTo) value;
+        else
+            return IntTo.max;
+    }
+    else
+    {
+        static assert(isUnsigned!IntFrom && isSigned!IntTo);
+
+        if (value < cast(Unsigned!IntTo) IntTo.max)
+            return cast(IntTo) value;
+        else
+            return IntTo.max;
+    }
+}
+
+///
+unittest
+{
+    assert((0).boundedConvert!uint == 0u);
+    assert((42).boundedConvert!uint == 42u);
+    assert((int.max).boundedConvert!uint == cast(uint) int.max);
+    assert((-1).boundedConvert!uint == 0u);
+}
+
+unittest
+{
+    import std.meta;
+
+    alias IntTypes = AliasSeq!(
+        byte,
+        ubyte,
+        short,
+        ushort,
+        int,
+        uint,
+        long,
+        ulong,
+        //cent,
+        //ucent,
+    );
+
+    static foreach (alias IntFrom; IntTypes)
+        static foreach (alias IntTo; IntTypes)
+        {
+            assert(IntFrom(0).boundedConvert!IntTo == IntTo(0));
+            assert(IntFrom(42).boundedConvert!IntTo == IntTo(42));
+        }
+
+    // IntFrom = byte
+    assert(boundedConvert!byte      (byte.max) == byte.max);
+    assert(boundedConvert!ubyte     (byte.max) == byte.max);
+    assert(boundedConvert!short     (byte.max) == byte.max);
+    assert(boundedConvert!ushort    (byte.max) == byte.max);
+    assert(boundedConvert!int       (byte.max) == byte.max);
+    assert(boundedConvert!uint      (byte.max) == byte.max);
+    assert(boundedConvert!long      (byte.max) == byte.max);
+    assert(boundedConvert!ulong     (byte.max) == byte.max);
+    assert(boundedConvert!byte      (byte.min) == byte.min);
+    assert(boundedConvert!ubyte     (byte.min) == ubyte.min);
+    assert(boundedConvert!short     (byte.min) == byte.min);
+    assert(boundedConvert!ushort    (byte.min) == ushort.min);
+    assert(boundedConvert!int       (byte.min) == byte.min);
+    assert(boundedConvert!uint      (byte.min) == uint.min);
+    assert(boundedConvert!long      (byte.min) == byte.min);
+    assert(boundedConvert!ulong     (byte.min) == ulong.min);
+
+    // IntFrom = ubyte
+    assert(boundedConvert!byte      (ubyte.max) == byte.max);
+    assert(boundedConvert!ubyte     (ubyte.max) == ubyte.max);
+    assert(boundedConvert!short     (ubyte.max) == ubyte.max);
+    assert(boundedConvert!ushort    (ubyte.max) == ubyte.max);
+    assert(boundedConvert!int       (ubyte.max) == ubyte.max);
+    assert(boundedConvert!uint      (ubyte.max) == ubyte.max);
+    assert(boundedConvert!long      (ubyte.max) == ubyte.max);
+    assert(boundedConvert!ulong     (ubyte.max) == ubyte.max);
+
+    // IntFrom = int
+    assert(boundedConvert!byte      (int.max) == byte.max);
+    assert(boundedConvert!ubyte     (int.max) == ubyte.max);
+    assert(boundedConvert!short     (int.max) == short.max);
+    assert(boundedConvert!ushort    (int.max) == ushort.max);
+    assert(boundedConvert!int       (int.max) == int.max);
+    assert(boundedConvert!uint      (int.max) == int.max);
+    assert(boundedConvert!long      (int.max) == int.max);
+    assert(boundedConvert!ulong     (int.max) == int.max);
+    assert(boundedConvert!byte      (int.min) == byte.min);
+    assert(boundedConvert!ubyte     (int.min) == ubyte.min);
+    assert(boundedConvert!short     (int.min) == short.min);
+    assert(boundedConvert!ushort    (int.min) == ushort.min);
+    assert(boundedConvert!int       (int.min) == int.min);
+    assert(boundedConvert!uint      (int.min) == uint.min);
+    assert(boundedConvert!long      (int.min) == int.min);
+    assert(boundedConvert!ulong     (int.min) == uint.min);
+
+    // IntFrom = uint
+    assert(boundedConvert!byte      (uint.max) == byte.max);
+    assert(boundedConvert!ubyte     (uint.max) == ubyte.max);
+    assert(boundedConvert!short     (uint.max) == short.max);
+    assert(boundedConvert!ushort    (uint.max) == ushort.max);
+    assert(boundedConvert!int       (uint.max) == int.max);
+    assert(boundedConvert!uint      (uint.max) == uint.max);
+    assert(boundedConvert!long      (uint.max) == uint.max);
+    assert(boundedConvert!ulong     (uint.max) == uint.max);
+}
+
+
 class EdgeExistsException : Exception
 {
     pure nothrow @nogc @safe this(
@@ -1296,10 +1425,11 @@ struct Graph(Node, Weight = void, Flag!"isDirected" isDirected = No.isDirected, 
                 }
             }
 
-            ref inout(Edge[]) opIndex(in Node node) inout
-            {
-                return incidentEdges[graph.indexOf(node)];
-            }
+            static if (!is(Node == size_t))
+                ref inout(Edge[]) opIndex(in Node node) inout
+                {
+                    return incidentEdges[graph.indexOf(node)];
+                }
 
             ref inout(Edge[]) opIndex(in size_t nodeIdx) inout
             {
@@ -1431,10 +1561,11 @@ struct Graph(Node, Weight = void, Flag!"isDirected" isDirected = No.isDirected, 
                 }
             }
 
-            size_t opIndex(in Node node) const
-            {
-                return degrees[graph.indexOf(node)];
-            }
+            static if (!is(Node == size_t))
+                size_t opIndex(in Node node) const
+                {
+                    return degrees[graph.indexOf(node)];
+                }
 
             size_t opIndex(in size_t nodeIdx) const
             {
@@ -3978,4 +4109,54 @@ unittest
         1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0,
         0, 0, 0, 1, 0,
     ]));
+}
+
+
+/// Encode/decode a pair of integers in a single integer. Caution: the encoded
+/// value is in Θ(a * b).
+Int encodePair(Int)(Int a, Int b) if (isIntegral!Int)
+{
+    // formula taken from https://mathforum.org/library/drmath/view/56036.html
+    return ((a + b)^^2 + 3*a + b) / 2;
+}
+
+/// ditto
+Int[2] decodePair(Int)(Int n) if (isIntegral!Int)
+{
+    import std.math : floor;
+
+    // formulas from https://mathforum.org/library/drmath/view/56036.html
+    Int c = cast(Int) floor((sqrt(8.0*n + 1.0) - 1.0)/2.0);
+    Int a = n - c*(c + 1)/2;
+    Int b = c - a;
+
+    return [a, b];
+}
+
+///
+unittest
+{
+    auto encoded = encodePair(42, 1337);
+
+    assert(encoded == 951552);
+
+    auto decoded = decodePair(encoded);
+
+    assert(decoded[0] == 42);
+    assert(decoded[1] == 1337);
+}
+
+unittest
+{
+    foreach (a; 0 .. 50)
+        foreach (b; 0 .. 50)
+        {
+            auto n = encodePair(a, b);
+            auto decoded = decodePair(n);
+            auto n2 = encodePair(decoded[0], decoded[1]);
+
+            assert(a == decoded[0]);
+            assert(b == decoded[1]);
+            assert(n == n2);
+        }
 }